requirements.txt
google-auth
google-cloud-compute
paramiko
google-api-python-client